Justeringer etter QA
