+++ src/frontend/komponenter/Fagsak/Vedtak/FeilutbetaltValuta/useFeilutbetaltValuta.tsx
@@ -1,5 +1,4 @@
-import type { FeltState } from '@navikt/familie-skjema';
-import { feil, ok, useFelt, useSkjema } from '@navikt/familie-skjema';
+import { useFelt, useSkjema } from '@navikt/familie-skjema';
 import type { Ressurs } from '@navikt/familie-typer';
 import { RessursStatus } from '@navikt/familie-typer';
 
@@ -11,74 +10,13 @@ import type {
     IRestFeilutbetaltValuta,
 } from '../../../../typer/eøs-feilutbetalt-valuta';
 import type { FamilieIsoDate } from '../../../../utils/kalender';
-import {
-    erIsoStringGyldig,
-    kalenderDato,
-    sisteDagIInneværendeMåned,
-} from '../../../../utils/kalender';
-import {
-    erFør,
-    kalenderDatoMedFallback,
-    TIDENES_ENDE,
-    TIDENES_MORGEN,
-} from '../../../../utils/kalender';
-import { erPositivtHeltall } from '../../../../utils/validators';
+import { validerFeilutbetaltBeløp, validerTom, validerFom } from './FeilutbetaltValutaUtil';
 
 interface IProps {
     behandlingId: number;
     feilutbetaltValuta?: IRestFeilutbetaltValuta;
     settFeilmelding: (feilmelding: string) => void;
 }
-
-const datoErFremITid = (dato: FamilieIsoDate): boolean => {
-    const nå = sisteDagIInneværendeMåned();
-
-    return erFør(nå, kalenderDato(dato));
-};
-
-const validerFom = (felt: FeltState<FamilieIsoDate>) => {
-    const fom = felt.verdi;
-
-    if (fom === '') return feil(felt, 'Du må velge en f.o.m-dato');
-    if (!erIsoStringGyldig(fom)) {
-        return feil(felt, 'Du må velge en gyldig f.o.m-dato');
-    }
-    if (datoErFremITid(fom)) {
-        return feil(felt, 'F.o.m kan ikke være senere enn inneværende måned');
-    }
-    return ok(felt);
-};
-
-const validerTom = (felt: FeltState<FamilieIsoDate>, fom: FamilieIsoDate) => {
-    const tom = felt.verdi;
-
-    if (tom === '') return feil(felt, 'Du må velge en t.o.m-dato');
-    if (!erIsoStringGyldig(tom)) {
-        return feil(felt, 'Du må velge en gyldig t.o.m-dato');
-    }
-    if (datoErFremITid(tom)) {
-        return feil(felt, 'T.o.m. kan ikke være senere enn inneværende måned');
-    }
-
-    const fomKalenderDato = kalenderDatoMedFallback(fom, TIDENES_MORGEN);
-    const tomKalenderDato = kalenderDatoMedFallback(tom, TIDENES_ENDE);
-    const fomDatoErFørTomDato = erFør(fomKalenderDato, tomKalenderDato);
-
-    if (!fomDatoErFørTomDato) {
-        return feil(felt, 'T.o.m. må være senere enn f.o.m');
-    }
-
-    return ok(felt);
-};
-
-const validerFeilutbetaltBeløp = (felt: FeltState<string>) => {
-    if (felt.verdi === '') {
-        return feil(felt, 'Beløp er påkrevd');
-    } else if (!erPositivtHeltall(felt.verdi)) {
-        return feil(felt, 'Feil format. Skriv inn et gyldig siffer.');
-    }
-    return ok(felt);
-};
 
 const useFeilutbetaltValuta = ({ feilutbetaltValuta, settFeilmelding, behandlingId }: IProps) => {
     const { settÅpenBehandling } = useBehandling();
